+++ lib/Activity/SettingData.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Data Analytics
+ * Analytics
  *
  * This file is licensed under the Affero General Public License version 3 or
  * later. See the LICENSE.md file.
@@ -43,7 +43,7 @@ class SettingData implements \OCP\Activity\ISetting
      */
     public function getName()
     {
-        return $this->l->t('<strong>Data Analytics</strong>: New data was added to a report');
+        return $this->l->t('<strong>Analytics</strong>: New data was added to a report');
     }
 
     /**