+++ lib/DataSession.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 /**
- * Data Analytics
+ * Analytics
  *
  * This file is licensed under the Affero General Public License version 3 or
  * later. See the LICENSE.md file.